+++ CHANGELOG.md
@@ -10,9 +10,9 @@ Tantivy 0.14.0
 - Removed dependency on `notify` #924. Replaced with `FileWatcher` struct that polls meta file every 500ms in background thread. (@halvorboe @guilload)
 - Added `FilterCollector`, which wraps another collector and filters docs using a predicate over a fast field (@barrotsteindev)
 - Simplified the encoding of the skip reader struct. BlockWAND max tf is now encoded over a single byte. (@pmasurel)
+- `FilterCollector` now supports all Fast Field value types (@barrotsteindev)
 
 This version breaks compatibility and requires users to reindex everything.
-
 
 Tantivy 0.13.2
 ===================